% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtl.R
\name{dbtl}
\alias{dbtl}
\title{Bradley-Terry-Luce density function}
\usage{
dbtl(Pi, worth, log = FALSE, Pi_full = NULL)
}
\arguments{
\item{Pi}{Matrix of partial or complete rankings, one row per ranking.}

\item{worth}{Vector of object worth parameters from a Bradley-Terry-Luce type distribution.}

\item{log}{Boolean indicating if loglikelihood should be returned.}

\item{Pi_full}{Matrix of objects considered when formulating each ranking (to account for groupwise comparisons). Should have same number of rows as Pi, and first entries of each row should match Pi. For example, if a judge considers obejcts 1,2,3,4 and provides the top-2 ranking 1<2, then Pi should have a row with entries 1,2 and Pi_full should have a row with entries 1,2,3,4.}
}
\value{
(Log) likelihood of rankings under a Bradley-Terry-Luce distribution.
}
\description{
This function calculates the density of observation(s) under a Bradley-Terry-Luce distribution.
}
\examples{
Pi <- matrix(c(1,2,3,4,5,2,1,NA,NA,NA),byrow=TRUE,nrow=2)
Pi_full <- matrix(c(1,2,3,4,5,2,1,3,4,NA),byrow=TRUE,nrow=2)
dbtl(Pi=Pi,worth=c(5,4,3,2,1),log=TRUE)
dbtl(Pi=Pi,worth=c(5,4,3,2,1),log=TRUE,Pi_full=Pi_full)
 
}
