% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtlb.R
\name{dbtlb}
\alias{dbtlb}
\title{BTL-Binomial density function}
\usage{
dbtlb(Pi, X, p, theta, M, log = FALSE, Pi_full = NULL)
}
\arguments{
\item{Pi}{Matrix of partial or complete rankings, one row per ranking.}

\item{X}{Matrix of ratings, one row per judge and one column per object.}

\item{p}{Vector of object qualities.}

\item{theta}{Numeric specifying the consensus strength.}

\item{M}{Numeric specifying maximum (=worst quality) integer rating.}

\item{log}{Boolean indicating if loglikelihood should be returned.}

\item{Pi_full}{Matrix of objects considered when formulating each ranking (to account for groupwise comparisons). Should have same number of rows as Pi, and first entries of each row should match Pi. For example, if a judge considers obejcts 1,2,3,4 and provides the top-2 ranking 1<2, then Pi should have a row with entries 1,2 and Pi_full should have a row with entries 1,2,3,4.}
}
\value{
(Log) likelihood of rankings and ratings under a BTL-B distribution.
}
\description{
This function calculates the density of observation(s) under a Bradley-Terry-Luce-Binomial (BTL-B) distribution.
}
\examples{
Pi <- matrix(c(1,2,3,4,5,2,1,NA,NA,NA),byrow=TRUE,nrow=2)
Pi_full <- matrix(c(1,2,3,4,5,2,1,3,4,NA),byrow=TRUE,nrow=2)
X <- matrix(c(0,1,2,3,4,1,2,2,5,5),byrow=TRUE,nrow=2)
p <- c(.1,.2,.5,.7,.8)
theta <- 10
M <- 6
dbtlb(Pi,X,p,theta,M,log=TRUE)
dbtlb(Pi,X,p,theta,M,log=TRUE,Pi_full)
 
}
