% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmb.R
\name{dmb}
\alias{dmb}
\title{Mallows-Binomial density function}
\usage{
dmb(rankings, ratings, p, theta, M, pi0 = NULL, log = FALSE)
}
\arguments{
\item{rankings}{A vector or matrix of rankings, potentially with attribute "assignments" to signify separate reviewer assignments. One rankings per row.}

\item{ratings}{Vector or matrix of ratings, with one row per ratings for each judge.}

\item{p}{Vector of object qualities.}

\item{theta}{Numeric specifying the Mallows scale parameter.}

\item{M}{Numeric specifying maximum (=worst quality) integer rating.}

\item{pi0}{Vector specifying the consensus (modal probability) ranking (useful to break ties in p).}

\item{log}{Boolean indicating if loglikelihood should be returned.}
}
\value{
(Log) likelihood of rankings under a Mallows-Binomial distribution.
}
\description{
This function calculates the density of observation(s) under a Mallows-Binomial distribution.
}
\examples{
rankings <- matrix(c(1,2,3,1,2,3),nrow=2,byrow=TRUE)
ratings <- matrix(c(2,4,6,4,2,6),nrow=2,byrow=TRUE)
dmb(rankings=c(1,2,3),ratings=c(2,4,6),p=c(.25,.5,.75),theta=2,M=8)
dmb(rankings=rankings,ratings=ratings,p=c(0.33,0.33,0.75),theta=6,M=8,pi0=c(1,2,3),log=TRUE)

}
