% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mb.R
\name{fit_mb}
\alias{fit_mb}
\title{Calculate the exact or approximate MLE of a Mallows-Binomial distribution using various methods}
\usage{
fit_mb(
  rankings,
  ratings,
  M,
  method = c("ASTAR", "Greedy", "GreedyLocal", "FV")
)
}
\arguments{
\item{rankings}{A matrix of rankings, potentially with attribute "assignments" to signify separate reviewer assignments. One ranking per row.}

\item{ratings}{A matrix of ratings, one row per judge and one column per object.}

\item{M}{Numeric specifying maximum (=worst quality) integer rating.}

\item{method}{String indicating which estimation method to use when estimating parameters. Allowable options are currently "ASTAR", "Greedy", "GreedyLocal", and "FV". Defaults to exact search, "ASTAR".}
}
\value{
List with elements pi0 (estimated consensus ranking MLE),  p (estimated object quality parameter MLE), theta (estimated scale parameter MLE), and numnodes (number of nodes traversed during algorithm, a measure of computational complexity).
}
\description{
This function calculates the exact or approximate MLE of a Mallows-Binomial distribution using a user-specified method.
}
\examples{
rankings <- matrix(c(1,2,3,4,2,1,NA,NA),byrow=TRUE,nrow=2)
ratings <- matrix(c(0,1,2,3,1,2,2,5),byrow=TRUE,nrow=2)
fit_mb(rankings=rankings,ratings=ratings,M=5,method="ASTAR")
fit_mb(rankings=rankings,ratings=ratings,M=5,method="Greedy")
fit_mb(rankings=rankings,ratings=ratings,M=5,method="GreedyLocal")
fit_mb(rankings=rankings,ratings=ratings,M=5,method="FV")

}
