% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mbmm.R
\name{fit_mbmm}
\alias{fit_mbmm}
\title{MLE and MAP Estimation of a MB Mixture Model model}
\usage{
fit_mbmm(
  Pi,
  X,
  M,
  I,
  J,
  Pi_full = NULL,
  tol = 1,
  maxiter = 50,
  scope = max(10, log(factorial(J))),
  verbose = FALSE,
  K = 1,
  alpha0 = 1,
  a = 1,
  b = 1,
  gamma1 = 1,
  gamma2 = 0
)
}
\arguments{
\item{Pi}{Matrix of partial or complete rankings, one row per ranking.}

\item{X}{Matrix of ratings, one row per judge and one column per object.}

\item{M}{Numeric specifying maximum (=worst quality) integer rating.}

\item{I}{Numeric specifying number of judges}

\item{J}{Numeric specifying number of objects}

\item{Pi_full}{Matrix of objects considered when formulating each ranking (to account for limited object access). Should have same number of rows as Pi, and first entries of each row should match Pi. For example, if a judge considers obejcts 1,2,3,4 and provides the top-2 ranking 1<2, then Pi should have a row with entries 1,2 and Pi_full should have a row with entries 1,2,3,4.}

\item{tol}{Numeric specifying stopping tolerance of loglikelihood in successive EM iterations.}

\item{maxiter}{Maximum number of EM iterations before returning results (default=50).}

\item{scope}{Number of rankings to sample at each update stage in the EM algorithm. When scope is large, algorithm will be slower but more likely to obtain an exact solution.}

\item{verbose}{Boolean specifying if iteration information should be printed while function runs.}

\item{K}{Integer specifying number of latent classes (default = 1, i.e., no latent classes)}

\item{alpha0}{Numeric or vector specifying hyperparameters of Dirichlet prior on latent classes. If K>1 and alpha0 is a single number, alpha0 becomes rep(alpha0,K). Default is a flat (uninformative) prior that corresponds to MLE.}

\item{a}{Numeric specifying first hyperparameter on Beta(a,b) prior on each p_jk. Default is a flat (uninformative) prior that corresponds to MLE.}

\item{b}{Numeric specifying second hyperparameter on Beta(a,b) prior on each p_jk. Default is a flat (uninformative) prior that corresponds to MLE.}

\item{gamma1}{Numeric specifying first hyperparameter on Gamma(gamma1,gamma2) prior on each theta_k. Default is a flat (uninformative) prior that corresponds to MLE.}

\item{gamma2}{Numeric specifying second hyperparameter on Gamma(gamma1,gamma2) prior on each theta_k. Default is a flat (uninformative) prior that corresponds to MLE.}
}
\value{
List of estimated parameters (alpha,p,pi0,theta,Z).
}
\description{
This function estimates the maximum likelihood or maximum a posteriori estimates of a Mallows-Binomial distribution with potentially K mixture components via an Expectation-Maximization (EM) algorithm.
}
\examples{
Pi <- matrix(c(1,2,3,4,5,2,1,NA,NA,NA),byrow=TRUE,nrow=2)
Pi_full <- matrix(c(1,2,3,4,5,2,1,3,4,NA),byrow=TRUE,nrow=2)
X <- matrix(c(0,1,2,3,4,1,2,2,5,5),byrow=TRUE,nrow=2)
fit_mbmm(Pi=Pi,X=X,M=6,I=2,J=5,tol=0.01)
fit_mbmm(Pi=Pi,X=X,M=6,I=2,J=5,tol=0.01,a=5,b=1,gamma1=10,gamma2=10)
 
}
