% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_heuristic.R
\name{lp_heuristic}
\alias{lp_heuristic}
\title{Calculate the LP heuristic of a Mallows-Binomial model}
\usage{
lp_heuristic(Q, Pi, I, J, order)
}
\arguments{
\item{Q}{Matrix of dimension J x J.}

\item{Pi}{Matrix of partial or complete rankings, one row per ranking.}

\item{I}{Numeric specifying number of judges}

\item{J}{Numeric specifying number of objects}

\item{order}{Vector specifying a top-r or complete ordering of the desired p vector.}
}
\value{
Numeric specifying the LP heuristic.
}
\description{
This function calculates the LP heuristic of a Mallows-Binomial model, for use during an A* tree search for the MLE of a Mallows-Binomial model.
}
