% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mcmc.R
\name{make_mcmc}
\alias{make_mcmc}
\title{Convert Posteriors Samples from fit-btlbmm function into (Split) MCMC Object}
\usage{
make_mcmc(gibbs, split = 1)
}
\arguments{
\item{gibbs}{Output of the fit_btlbmm function.}

\item{split}{Number of roughly-equal sized splits to perform of the posterior. Helpful if looking to calculate the Gelman diagnostic on the posterior. Default is no splitting (split=1).}
}
\value{
MCMC list object.
}
\description{
This function samples approximately from the posterior of a Bradley-Terry-Luce-Binomial distribution with potentially K mixture components via an adaptive Metropolis-Hastings-within-Gibbs algorithm.
}
