% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{Psi function}
\usage{
psi(theta, J, R, log = FALSE)
}
\arguments{
\item{theta}{Non-negative scale parameter.}

\item{J}{Vector of positive integers indicating total number of objects. If length(J)>1, R must be of same length or a single value.}

\item{R}{Vector of positive integer <=\code{J} indicating size of partial ranking. If length(R)>1, J must be of same length or a single value.}

\item{log}{Boolean indicating if log(psi(theta)) should be returned.}
}
\value{
Numeric vector representing normalizing constant of a Mallows distribution.
}
\description{
This function calculates the normalizing constant of a Mallows distribution under the Kendall distance
}
\examples{
psi(theta=1,J=10,R=8)
psi(theta=2,J=c(4,4,3),R=c(2,2,1),log=TRUE)

}
