% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_conditional.R
\name{theta_conditional}
\alias{theta_conditional}
\title{Estimate theta in a Mallows-Binomial distribution under a constraint.}
\usage{
theta_conditional(rankings, D, J)
}
\arguments{
\item{rankings}{A matrix of rankings, potentially with attribute "assignments" to signify separate reviewer assignments. One ranking per row.}

\item{D}{Numeric specifying a minimum total kendall distance between the observed rankings and a supposed central ranking.}

\item{J}{Numeric specifying the total number of objects to be assessed.}
}
\value{
Numeric specifying the MLE of theta under a constraint.
}
\description{
This function calculates the MLE of the consensus scale parameter, theta, in a Mallows-Binomial(p,theta) distribution given a constraint.
}
\examples{
rankings <- matrix(c(1,2,3,4,2,1,NA,NA),byrow=TRUE,nrow=2)
theta_conditional(rankings,D=1,J=4)
theta_conditional(rankings,D=2,J=4)

}
