% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rankings.R
\name{to_rankings}
\alias{to_rankings}
\title{Convert ranks into rankings (orderings)}
\usage{
to_rankings(ranks, assignments = NULL, rankings = NULL)
}
\arguments{
\item{ranks}{Matrix or vector of ranks, such that the (i,j) entry includes the rank given by judge i to proposal j. NA used to indicate no rank was assigned to a proposal, which may occur for two reasons: (1) If "assignments" is not specified or the (i,j) entry of assignments is TRUE, NA indicates a proposal was considered worse than all ranked proposals. If the (i,j) entry of assignments is FALSE, NA indicates a proposal simply was not considered by the judge and no information can be gleaned from the missing rank.}

\item{assignments}{Matrix of booleans, such that the (i,j) entry is TRUE if judge i was assigned to review proposal j, and FALSE otherwise. If assignments is NULL, we assume all judges considered all proposals.}

\item{rankings}{Matrix or vector of rankings, with one row per judge. If rankings is specified, the function is used only to add the assignments attribute to the rankings file; any ranks input is ignored.}
}
\value{
Matrix of rankings, one row per ranking. If "assignments" argument is specified, then the rankings matrix will have the attribute "assignments".
}
\description{
This function converts a matrix of ranks into a matrix of orderings, potentially including reviewer assignments as an attribute of the ranking matrix.
}
\examples{
ranks <- matrix(data=c(4,2,3,1,NA,1,2,3,NA,NA,1,NA),byrow=TRUE,nrow=3)
assignments=matrix(TRUE,byrow=TRUE,nrow=3,ncol=4)
to_rankings(ranks=ranks)
to_rankings(ranks=ranks,assignments=assignments)
to_rankings(assignments=matrix(TRUE,nrow=1,ncol=3),rankings=c(3,2,1))

}
