% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalcostheuristic_MB.R
\name{totalcostheuristic_MB}
\alias{totalcostheuristic_MB}
\title{Calculate the naive total cost heuristic of a Mallows-Binomial model}
\usage{
totalcostheuristic_MB(Q, rankings, ratings, M, order)
}
\arguments{
\item{Q}{Matrix of dimension J x J.}

\item{rankings}{A matrix of rankings, potentially with attribute "assignments" to signify separate reviewer assignments. One ranking per row.}

\item{ratings}{A matrix of ratings, one row per judge and one column per object.}

\item{M}{Numeric specifying maximum (=worst quality) integer rating.}

\item{order}{Vector specifying a top-r or complete ordering of the desired p vector.}
}
\value{
List of three numeric values: estimated object quality vector (phat), estimated consensus scale parameter (thetahat), and the total cost heuristic (totalcostheuristic)
}
\description{
This function calculates the total cost heuristic of a Mallows-Binomial model given Q, rankings, ratings, M, and an order, for use during an A* tree search for the MLE of a Mallows-Binomial model.
}
\examples{
Q <- matrix(c(0,.5,0,0,.5,0,0,0,1,1,0,0,1,1,.5,0),nrow=4,ncol=4)
rankings <- matrix(c(1,2,3,4,2,1,NA,NA),byrow=TRUE,nrow=2)
ratings <- matrix(c(0,1,2,3,1,2,2,5),byrow=TRUE,nrow=2)
totalcostheuristic_MB(Q=Q,rankings=rankings,ratings=ratings,M=5,order=c(1,2))
totalcostheuristic_MB(Q=Q,rankings=rankings,ratings=ratings,M=5,order=c(2,1,4,3))

}
