% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_mb.R
\name{ci_mb}
\alias{ci_mb}
\title{Calculate Confidence Intervals for Mallows-Binomial parameters.}
\usage{
ci_mb(
  rankings,
  ratings,
  M,
  interval = 0.9,
  nsamples = 10,
  all = FALSE,
  method = "ASTAR"
)
}
\arguments{
\item{rankings}{A matrix of rankings, potentially with attribute "assignments" to signify separate reviewer assignments. One ranking per row.}

\item{ratings}{A matrix of ratings, one row per judge and one column per object.}

\item{M}{Numeric specifying maximum (=worst quality) integer rating.}

\item{interval}{Numeric between 0 and 1 specifying the confidence interval (e.g., .90 indicates a 90\% confidence interval). Defaults to 0.90.}

\item{nsamples}{Numeric indicating desired number of bootstrap samples to be used when calculating confidence intervals. Defaults to 100.}

\item{all}{Boolean indicating if estimated parameters from all bootstrap samples should be returned. Defaults to FALSE.}

\item{method}{String indicating which estimation method to use when estimating parameters. Allowable options are currently "ASTAR", "Greedy", "GreedyLocal", and "FV". Defaults to exact search, "ASTAR".}
}
\value{
List with elements ci (matrix of confidence intervals for Mallows-Binomial parameters), ci_ranks (matrix of confidence intervals for object ranks), bootstrap_pi0 (matrix of bootstrap consensus rankings; returned only if all=TRUE), and bootstrap_ptheta (matrix of bootstrap estimates of (p,theta); returned only if all=TRUE).
}
\description{
This function calculates confidence intervals for parameters in a Mallows-Binomial model using the nonparametric bootstrap.
}
\examples{
rankings <- matrix(c(1,2,3,4,2,1,NA,NA),byrow=TRUE,nrow=2)
ratings <- matrix(c(0,1,2,3,1,2,2,5),byrow=TRUE,nrow=2)
ci_mb(rankings=rankings,ratings=ratings,M=5,method="FV",all=TRUE)

}
