% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQ.R
\name{getQ}
\alias{getQ}
\title{Calculate Q Matrix}
\usage{
getQ(Pi, J)
}
\arguments{
\item{Pi}{Matrix of partial or complete rankings, one row per ranking.}

\item{J}{Total number of objects assessed.}
}
\value{
Matrix Q of dimensions J x J.
}
\description{
This function calculates the Q matrix given a collection of (partial) rankings.
}
\examples{
Pi <- matrix(c(1,2,3,4,2,1,NA,NA),byrow=TRUE,nrow=2)
getQ(Pi=Pi,J=4)

}
