% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreedyLocal.R
\name{GreedyLocal}
\alias{GreedyLocal}
\title{Estimate the MLE of a Mallows-Binomial distribution using the GreedyLocal method}
\usage{
GreedyLocal(rankings, ratings, M)
}
\arguments{
\item{rankings}{A matrix of rankings, potentially with attribute "assignments" to signify separate reviewer assignments. One ranking per row.}

\item{ratings}{A matrix of ratings, one row per judge and one column per object.}

\item{M}{Numeric specifying maximum (=worst quality) integer rating.}
}
\value{
List with elements pi0 (estimated consensus ranking MLE),  p (estimated object quality parameter MLE), theta (estimated scale parameter MLE), and numnodes (number of nodes traversed during algorithm, a measure of computational complexity).
}
\description{
This function estimates the MLE of a Mallows-Binomial distribution using the GreedyLocal method, which is identical to the Greedy method but includes an automatic and targeted post-hoc local search.
}
\examples{
data("ToyData1")
GreedyLocal(ToyData1$rankings,ToyData1$ratings,ToyData1$M)

}
