% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall.R
\name{kendall}
\alias{kendall}
\title{Kendall's tau function}
\usage{
kendall(pi, pi0)
}
\arguments{
\item{pi}{A partial or complete ranking.}

\item{pi0}{A complete ranking.}
}
\value{
Numeric Kendall's tau distance between pi and pi0.
}
\description{
This function calculates Kendall's tau distance between a (partial) ranking pi and a complete ranking pi0.
}
\examples{
kendall(pi=c(1,2,3),pi0=c(4,3,2,1))
kendall(pi=c(1,2,3),pi0=c(1,3,2))
 
}
